requirements.txt
aiogram==2.24
requests==2.28.2
pip~=23.3.1
python-dotenv~=1.0.0